backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
psycopg[binary]>=3.2.0
python-dotenv==1.0.1
waitress>=3.0.0
